fix: 결제 웹훅 데이터 무결성 및 보안 강화 완료 (Phase 3+4/4)
## 주요 변경사항

### Phase 3: 데이터 무결성 강화 ✅

**1. amount Float → Integer 변경** (Line 25 of subscription.py)
- **app/models/subscription.py**
  - `Column(Float)` → `Column(Integer)`
  - 부동소수점 오류 방지 (0.1 + 0.2 ≠ 0.3 문제 해결)
  - 금액은 원 단위 정수로만 처리
- **migrations/versions/93bbca6dac25_**
  - SQLite batch_alter_table 사용
  - 테이블 재생성 방식으로 타입 변경
  - 마이그레이션 성공 확인 완료

**이슈**: Float 타입으로 인한 부동소수점 오류 위험
**해결**: Integer 타입 사용 → 정확한 금액 계산

### Phase 4: 보안 및 테스트 강화 ✅

**2. verify_webhook_signature 403 반환** (Line 75-77 of payment.py)
- **app/controllers/payment.py**
  - 서명 검증 실패 시 403 Forbidden 즉시 반환
  - 잘못된 웹훅 차단 (보안 강화)
  - HTTPException 처리 추가

**이슈**: 서명 검증 실패해도 200 OK 반환
**해결**: 403 Forbidden 반환 → 무단 웹훅 차단

**3. 추가 테스트 케이스 작성** (3개 신규 테스트)

**tests/unit/test_payment_webhook.py** (총 8개 테스트)

**신규 테스트 1: test_partial_canceled_status()**
```python
# PARTIAL_CANCELED 상태 처리 검증
# - 부분 취소 로그 기록 확인
# - 성공 응답 확인
```

**신규 테스트 2: test_user_not_found_pending_save()**
```python
# User 없을 때 pending 저장 검증
# - error 응답 확인
# - Payment pending 상태 저장 확인
# - user_id 저장 확인
```

**신규 테스트 3: test_amount_integer_type()**
```python
# amount Integer 타입 검증
# - Payment 레코드 생성 확인
# - amount 타입이 int인지 확인
# - amount 값 정확성 확인
```

## 수정된 버그

**🟢 LOW (3개 - 모두 수정)**
- ✅ amount Float → Integer 타입 변경
- ✅ 서명 검증 실패 시 403 반환
- ✅ 테스트 커버리지 확대 (5개 → 8개)

## 완료 통계

**전체 20개 이슈 수정 현황**:
- 🔴 **CRITICAL (7개)**: ✅ 100% 완료 (Phase 1)
- 🟡 **MEDIUM (5개)**: ✅ 100% 완료 (Phase 2)
- 🟢 **LOW (3개)**: ✅ 100% 완료 (Phase 3+4)
- 🔵 **INFO (5개)**: 문서화 및 개선 사항 (선택)

**총 15개 필수 이슈 수정 완료 (100%)**

## 테스트 결과
```
[SUCCESS] All payment webhook tests passed!

Processed Scenarios:
  [PASS] Webhook signature verification (security)
  [PASS] Payment completed (DONE)
  [PASS] Payment canceled (CANCELED)
  [PASS] Idempotency guarantee (prevent duplicates)
  [PASS] Retry logic
  [PASS] Partial canceled status (NEW)
  [PASS] User not found pending save (NEW)
  [PASS] Amount integer type validation (NEW)

Supported Payment Statuses:
  - READY (waiting)
  - IN_PROGRESS (processing)
  - DONE (completed)
  - CANCELED (canceled)
  - ABORTED (failed)
  - EXPIRED (expired)
  - PARTIAL_CANCELED (partial cancel) [NEW]
  - WAITING_FOR_DEPOSIT (virtual account waiting)
```

## 마이그레이션 히스토리
1. `dc29c046ce38`: payment_key 컬럼 추가 (Phase 1)
2. `93bbca6dac25`: amount Float → Integer 변경 (Phase 3)

## 최종 검증

**데이터 무결성**:
- ✅ amount Integer로 부동소수점 오류 방지
- ✅ payment_key로 취소 추적 가능
- ✅ 트랜잭션 롤백으로 원자성 보장

**보안**:
- ✅ 웹훅 서명 검증 (HMAC SHA256)
- ✅ 서명 실패 시 403 Forbidden
- ✅ 민감 정보 마스킹 (payment_key, order_id, amount)

**기능 완성도**:
- ✅ 8가지 결제 상태 처리
- ✅ 3가지 웹훅 이벤트 처리
- ✅ 멱등성 보장
- ✅ 재시도 시스템
- ✅ User 없을 때 pending 저장

**테스트 커버리지**:
- ✅ 8개 테스트 케이스 (5개 기본 + 3개 신규)
- ✅ 100% 테스트 통과

## 다음 단계

Phase 1-4 완료! 결제 웹훅 시스템 완성.

**선택 사항 (INFO 레벨)**:
- 추가 테스트 케이스 (DEPOSIT_CALLBACK, CANCEL_STATUS_CHANGED)
- 성능 최적화
- 추가 문서화

**다음 기능 (원래 로드맵)**:
- Day 4-5: 구독 자동 갱신 시스템

## 관련 문서
- claudedocs/PAYMENT_WEBHOOK_REVIEW.md (20개 이슈 전체)
- claudedocs/ERROR_TRACKING.md

+++ app/models/subscription.py
@@ -22,7 +22,7 @@ class Payment(Base):
     id = Column(Integer, primary_key=True, index=True)
     user_id = Column(Integer, ForeignKey("users.id"))
     subscription_id = Column(Integer, ForeignKey("subscriptions.id"))
-    amount = Column(Float, nullable=False)
+    amount = Column(Integer, nullable=False)  # 금액 (원 단위, Integer로 변경하여 부동소수점 오류 방지)
     currency = Column(String, default="KRW")
     status = Column(String, nullable=False)
     payment_gateway_charge_id = Column(String, unique=True)  # order_id (예: owl_123_1696305600_pro)
